Mise à jour des fichiers

+++ Code/SAE/PHP/Models/Model.php
@@ -28,14 +28,14 @@ class Model {
          $req = $this->bd->prepare('SELECT DISTINCT nconst FROM titleprincipals where tconst= :tconst');
          $req->bindValue(":tconst", $tconst);
          $req->execute();
-         return $req->fetchall();
+         return $req->fetchAll(PDO::FETCH_ASSOC);
     }
     
     public function getTitreByNconst($nconst) {
          $req = $this->bd->prepare('SELECT DISTINCT tconst FROM titleprincipals where nconst= :nconst');
          $req->bindValue(":nconst", $nconst);
          $req->execute();
-         return $req->fetchall();
+         return $req->fetchAll(PDO::FETCH_ASSOC);
     }
 
     public function getActorInformationByNconst($nconst) {
@@ -57,6 +57,13 @@ class Model {
         $req->bindValue(":tconst", $tconst);
         $req->execute();
         return $req->fetchall();
+    }
+
+    public function getTitreAndRatingInformationByTconst($tconst) {
+        $req = $this->bd->prepare("SELECT * FROM titlebasics LEFT JOIN titleratings ON titlebasics.tconst=titleratings.tconst where titlebasics.tconst= :tconst");
+        $req->bindValue(":tconst", $tconst);
+        $req->execute();
+        return $req->fetchAll(PDO::FETCH_ASSOC);
     }
 
 
@@ -94,17 +101,21 @@ class Model {
     }
 
     public function getOmdbPoster(String $tconst){
-        $getContentsOmdb = file_get_contents('http://www.omdbapi.com/?i=' . $tconst . '&plot=full&apikey=' . $this->omdbApi);
-        $omdb = json_decode($getContentsOmdb, TRUE);
-        
-        if ($omdb != null){
+        try{
+            $getContentsOmdb = file_get_contents('http://www.omdbapi.com/?i=' . $tconst . '&plot=full&apikey=' . $this->omdbApi);
+            $omdb = json_decode($getContentsOmdb, TRUE);
+            
+            if ($omdb != null){
 
-            if ($omdb["Poster"] == "N/A") {
-                return "Content/img/NoImageAvailable.png";
-            } else {
-                return $omdb["Poster"];
+                if ($omdb["Poster"] == "N/A") {
+                    return "Content/img/NoImageAvailable.png";
+                } else {
+                    return $omdb["Poster"];
+                }
+
             }
-
+        }catch (Exception $e) {
+            return "Content/img/NoImageAvailable.png";
         }
 
     }
@@ -112,48 +123,53 @@ class Model {
     
     //Wikipedia affiche
     public function getActorPoster(String $nconst){
-        
-        $id = $this->bd->prepare("SELECT nconst, lien FROM afficheacteurs WHERE nconst= :nconst");
-        $id->execute(['nconst' => $nconst]);
-            
-        if ($id->rowCount() > 0) {
-            $lien = $id->fetch();
-            if ($lien['lien'] != ""){
-                return $lien['lien'];
+        try{
+            $id = $this->bd->prepare("SELECT nconst, lien FROM afficheacteurs WHERE nconst= :nconst");
+            $id->execute(['nconst' => $nconst]);
+                
+            if ($id->rowCount() > 0) {
+                $lien = $id->fetch();
+                if ($lien['lien'] != ""){
+                    return $lien['lien'];
+                } else {
+                    return "Content/img/NoPictureAvailable.png";
+                }
             } else {
-                return "Content/img/NoPictureAvailable.png";
+                $pnconst = array("nconst" => $nconst);
+                $json = json_encode($pnconst);
+                file_put_contents("/home/DraCorporation/public_html/Content/json/" . $nconst . ".json", $json);
+
+                $command = "/usr/bin/python3 /home/DraCorporation/public_html/Content/json/python/AfficheActeurs.py 2>&1";
+
+                try {
+                    exec($command, $output, $status);
+                    /*if ($status !== 0) {
+                        echo "Erreur lors de l'exécution de la commande: $command";
+                        var_dump($output);
+                        exit();
+                    }//*/
+                    
+                } catch (Exception $e) {
+                    echo 'Erreur lors de l\'exécution du script Python : ',  $e->getMessage(), "\n";
+                }
+                if (file_exists("/home/DraCorporation/public_html/Content/json/". $nconst . '_resultat' . '.json')){    
+                    $reponse = file_get_contents("/home/DraCorporation/public_html/Content/json/". $nconst . '_resultat' . '.json');
+                    $resultat = str_replace(array('[', ']', '"'), '', $reponse);
+
+                    if(file_exists("/home/DraCorporation/public_html/Content/json/" . $nconst . ".json")){unlink("/home/DraCorporation/public_html/Content/json/" . $nconst . '.json');}
+                    if(file_exists("/home/DraCorporation/public_html/Content/json/" . $nconst . '_resultat' . '.json')){unlink("/home/DraCorporation/public_html/Content/json/" . $nconst . '_resultat' . '.json');}
+                } else{
+                    $resultat = "Content/img/NoPictureAvailable.png";
+                }
             }
-        } else {
-            $pnconst = array("nconst" => $nconst);
-            $json = json_encode($pnconst);
-            file_put_contents("/home/DraCorporation/public_html/Content/json/" . $nconst . ".json", $json);
-
-            $command = "/usr/bin/python3 /home/DraCorporation/public_html/Content/json/python/AfficheActeurs.py 2>&1";
-
-            try {
-                exec($command, $output, $status);
-                /*if ($status !== 0) {
-                    echo "Erreur lors de l'exécution de la commande: $command";
-                    var_dump($output);
-                    exit();
-                }//*/
-                
-             } catch (Exception $e) {
-                echo 'Erreur lors de l\'exécution du script Python : ',  $e->getMessage(), "\n";
-             }
-                
-            $reponse = file_get_contents("/home/DraCorporation/public_html/Content/json/". $nconst . '_resultat' . '.json');
-            $resultat = str_replace(array('[', ']', '"'), '', $reponse);
-
-            if(file_exists("/home/DraCorporation/public_html/Content/json/" . $nconst . ".json")){unlink("/home/DraCorporation/public_html/Content/json/" . $nconst . '.json');}
-            if(file_exists("/home/DraCorporation/public_html/Content/json/" . $nconst . '_resultat' . '.json')){unlink("/home/DraCorporation/public_html/Content/json/" . $nconst . '_resultat' . '.json');}        
-        
-
-            return $resultat;
+        }catch (Exception $e){
+            $resultat = "Content/img/NoPictureAvailable.png";
 
         }
-        
-    }
+
+        return $resultat;
+
+        }
 
 
     /**
@@ -169,7 +185,7 @@ class Model {
      }
 
      public function getPersonneInformation($nconst){
-        $req = $this->bd->prepare('SELECT  nconst, primaryname, birthyear, deathyear, primaryprofession FROM namebasics where nconst= :nconst');
+        $req = $this->bd->prepare('SELECT  nconst, primaryname, birthyear, deathyear, primaryprofession, knownfortitles FROM namebasics where nconst= :nconst');
         $req->bindValue("nconst",$nconst);
         $req->execute();
         return $req->fetchAll(PDO::FETCH_ASSOC);
@@ -184,7 +200,7 @@ class Model {
      }
 
      public function recherchePersonneAvecArgu($arguRecherche){
-        $queryPerso = "SELECT nconst, primaryname FROM namebasics WHERE similarity(lower(unaccent(primaryname)), lower(unaccent(:arg))) > 0.4";
+        $queryPerso = "SELECT nconst, primaryname FROM namebasics WHERE similarity(lower(unaccent(primaryname)), lower(unaccent(:arg))) > 0.4 ";
         $resultPerso = $this->bd->prepare($queryPerso);
         $resultPerso->bindValue(':arg', $this->bd->quote($arguRecherche));
         $resultPerso->execute();
